Rename namespace to match folder
