+++ Cpp2IL.Core/Graphs/Processors/IBlockProcessor.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Cpp2IL.Core.ISIL;
 
-namespace Cpp2IL.Core.Graphs.Intermediate
+namespace Cpp2IL.Core.Graphs.Processors
 {
     internal interface IBlockProcessor
     {
